RDND_proper/models/tapnet/tapnet/requirements_inference.txt
notebook
jupyter_http_over_ws
chex
optax
dm-haiku
dm-tree
typing_extensions
matplotlib
mediapy
opencv-python
einshape
tensorflow
tensorflow-datasets
